Example package rename add modify development config
