+++ engine/demo/src/main/java/pl/touk/nussknacker/engine/javademo/TransactionAmountAggregator.java
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.javaexample;
+package pl.touk.nussknacker.engine.javademo;
 
 import org.apache.flink.api.common.functions.RichMapFunction;
 import org.apache.flink.api.common.state.ValueState;
