+++ engine/demo/src/main/resources/META-INF/services/pl.touk.nussknacker.engine.api.process.ProcessConfigCreator
@@ -1,1 +1,1 @@
-pl.touk.nussknacker.engine.example.DemoProcessConfigCreator
+pl.touk.nussknacker.engine.demo.DemoProcessConfigCreator
