+++ engine/demo/src/main/scala/pl/touk/nussknacker/engine/demo/UtilProcessHelper.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.example
+package pl.touk.nussknacker.engine.demo
 
 import io.circe.Encoder
 import pl.touk.nussknacker.engine.api.{Documentation, ParamName}
