+++ ui/server/develConf/sample/application.conf
@@ -84,9 +84,9 @@ standaloneEngineProcessLocation: "/tmp/standaloneProcesses"
 attachmentsPath: "/tmp/attachments"
 
 categoriesConfig: {
-  "FraudDetection": "streaming",
-  "Recommendations": "streaming",
-  "Default": "streaming"
+  "FraudDetection": "streaming-demo",
+  "Recommendations": "streaming-demo",
+  "Default": "streaming-generic"
   "Category1": "streaming"
   "Category2": "streaming"
   "ServerRestApi": "request-response"
@@ -116,6 +116,24 @@ processTypes {
     }
     modelConfig = ${streamingModelConfig}
   }
+ "streaming-generic": {
+    engineConfig {
+      jobManagerTimeout: 1m
+      type: "flinkStreaming"
+      restUrl: ${flinkRestUrl}
+      queryableStateProxyUrl: ${flinkQueryableStateProxyUrl}
+    }
+    modelConfig = ${streamingModelGenericConfig}
+  }
+  "streaming-demo": {
+    engineConfig {
+      jobManagerTimeout: 1m
+      type: "flinkStreaming"
+      restUrl: ${flinkRestUrl}
+      queryableStateProxyUrl: ${flinkQueryableStateProxyUrl}
+    }
+    modelConfig = ${streamingModelDemoConfig}
+  }
   "request-response": {
     modelConfig {
       classPath: ["../../engine/standalone/engine/sample/target/scala-2.12/standaloneSample.jar"]
@@ -135,11 +153,52 @@ signalsTopic: ${?SIGNALS_TOPIC}
 kafkaAddress: "localhost:3032"
 kafkaAddress: ${?KAFKA_ADDRESS}
 
-streamingModel: "../../engine/flink/management/sample/target/scala-2.12/managementSample.jar"
-streamingModel: ${?STREAMING_MODEL}
-
 streamingModelConfig {
-  classPath: [${streamingModel}]
+  classPath: ["../../engine/flink/management/sample/target/scala-2.12/managementSample.jar"]
+  environment: "test"
+
+  signals {
+    topic: ${signalsTopic}
+  }
+
+  timeout: 10s
+  restartInterval: 10s
+  checkpointInterval: 10s
+
+  kafka {
+    kafkaAddress: ${kafkaAddress}
+  }
+
+  asyncExecutionConfig {
+    bufferSize: 50
+    parallelismMultiplier: 1
+  }
+}
+
+streamingModelGenericConfig {
+  classPath: ["../../engine/flink/generic/target/scala-2.12/genericModel.jar"]
+  environment: "test"
+
+  signals {
+    topic: ${signalsTopic}
+  }
+
+  timeout: 10s
+  restartInterval: 10s
+  checkpointInterval: 10s
+
+  kafka {
+    kafkaAddress: ${kafkaAddress}
+  }
+
+  asyncExecutionConfig {
+    bufferSize: 50
+    parallelismMultiplier: 1
+  }
+}
+
+streamingModelDemoConfig {
+  classPath: ["../../engine/demo/target/scala-2.12/demoModel.jar"]
   environment: "test"
 
   signals {
