+++ ui/server/develConf/sample/users.conf
@@ -29,17 +29,17 @@ rules: [
   {
     role: "Reader"
     permissions: ["Read"]
-    categories: ["Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
+    categories: ["ServerRestApi", "Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
   },
   {
     role: "Writer"
     permissions: ["Read", "Write"]
-    categories: ["Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
+    categories: ["ServerRestApi", "Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
   },
   {
     role: "Deployer"
     permissions: ["Read", "Write", "Deploy"]
     globalPermissions: ["AdminTab"]
-    categories: ["Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
+    categories: ["ServerRestApi", "Category1", "Category2", "FraudDetection", "Recommendations", "Default"]
   }
 ]
